Middleware issue fixed, Influcners routes protected, Models and Migrations added for Profile and  Post & Sotry along with their insigts migrations

+++ src/app/database/migrations/20251114070220_create_ad_campaign_demographics_age_ranges_table.js
@@ -3,8 +3,8 @@ module.exports = {
   async up(queryInterface, Sequelize) {
     await queryInterface.sequelize.query(`
       CREATE TABLE IF NOT EXISTS ad_campaign_demographic_age_ranges (
-        ad_campaign_demographic_id  integer REFERENCES ad_campaign_demographics (id),
-        age_range_id  integer REFERENCES ad_campaign_age_ranges (id)
+        ad_campaign_demographic_id  BIGINT REFERENCES ad_campaign_demographics (id),
+        age_range_id                INTEGER REFERENCES ad_campaign_age_ranges (id)
       )
     `);
   },
