+++ src/app/database/migrations/20251114070232_create_ad_campaign_locations_table.js
@@ -3,12 +3,12 @@ module.exports = {
     async up(queryInterface, Sequelize) {
         await queryInterface.sequelize.query(`
             CREATE TABLE IF NOT EXISTS ad_campaign_locations (
-                id              integer PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-                ad_campaign_id  integer REFERENCES ad_campaigns (id),
-                data_country_id integer NOT NULL REFERENCES data_countries (id),
-                data_state_id   integer NULL REFERENCES data_states (id),
-                data_city_id    integer NULL REFERENCES data_cities (id),
-                radius_miles    integer DEFAULT 5
+                id              BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+                ad_campaign_id  BIGINT REFERENCES ad_campaigns (id),
+                data_country_id INTEGER NOT NULL REFERENCES data_countries (id),
+                data_state_id   INTEGER NULL REFERENCES data_states (id),
+                data_city_id    INTEGER NULL REFERENCES data_cities (id),
+                radius_miles    INTEGER DEFAULT 5
             )
         `);
     },
